Propagated dice and partially dealt with int issue

+++ ninabaculinao.spritebuilder/Source/Gameplay.m
@@ -46,7 +46,7 @@
 
 - (void)step
 {
-    _scoreLabel.string = [NSString stringWithFormat:@"%d", _grid.score];
+    _scoreLabel.string = [NSString stringWithFormat:@"%ld", (long)_grid.score];
     [_grid spawnDice];
 }
